added component header
